Added pom.xml Added Op Player Only Fixed plugin.yml for demoplugin Code Cleanup
