--- src/resources/plugin.yml
@@ -1,3 +1,0 @@
-name: Demo
-main: net.lordsofcode.framework.example.DemoPlugin
-version: 1.0
